+++ standalone-metastore/metastore-common/src/main/java/org/apache/hadoop/hive/metastore/HiveMetaStoreClient.java
@@ -34,10 +34,6 @@ import org.apache.hadoop.hive.metastore.client.SynchronizedMetaStoreClientProxy;
 import org.apache.hadoop.hive.metastore.client.ThriftHiveMetaStoreClient;
 import org.apache.thrift.TException;
 
-import java.lang.reflect.InvocationHandler;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.lang.reflect.Proxy;
 import java.util.List;
 import java.util.Map;
 
@@ -77,8 +73,8 @@ public class HiveMetaStoreClient extends BaseMetaStoreClientProxy implements IMe
 
   private static IMetaStoreClient createUnderlyingClient(Configuration conf, HiveMetaHookLoader hookLoader,
       ThriftHiveMetaStoreClient thriftClient) {
-    IMetaStoreClient clientWithHook = new HookEnabledMetaStoreClientProxy(conf, hookLoader, thriftClient);
-    IMetaStoreClient synchronizedClient = new SynchronizedMetaStoreClientProxy(conf, clientWithHook);
+    IMetaStoreClient clientWithHook = HookEnabledMetaStoreClientProxy.newClient(conf, hookLoader, thriftClient);
+    IMetaStoreClient synchronizedClient = SynchronizedMetaStoreClientProxy.newClient(conf, clientWithHook);
     return synchronizedClient;
   }
 
@@ -183,37 +179,5 @@ public class HiveMetaStoreClient extends BaseMetaStoreClientProxy implements IMe
 
   public static String getProcessorIdentifier() {
     return ThriftHiveMetaStoreClient.getProcessorIdentifier();
-  }
-
-  /**
-   * Creates a synchronized wrapper for any {@link IMetaStoreClient}.
-   * This may be used by multi-threaded applications until we have
-   * fixed all reentrancy bugs.
-   *
-   * @param client unsynchronized client
-   * @return synchronized client
-   */
-  public static IMetaStoreClient newSynchronizedClient(IMetaStoreClient client) {
-    return (IMetaStoreClient) Proxy.newProxyInstance(
-        ThriftHiveMetaStoreClient.class.getClassLoader(),
-        new Class[]{IMetaStoreClient.class},
-        new SynchronizedHandler(client));
-  }
-
-  private static class SynchronizedHandler implements InvocationHandler {
-    private final IMetaStoreClient client;
-
-    SynchronizedHandler(IMetaStoreClient client) {
-      this.client = client;
-    }
-
-    @Override
-    public synchronized Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
-      try {
-        return method.invoke(client, args);
-      } catch (InvocationTargetException e) {
-        throw e.getTargetException();
-      }
-    }
   }
 }
