+++ standalone-metastore/metastore-common/src/main/java/org/apache/hadoop/hive/metastore/client/SynchronizedMetaStoreClientProxy.java
@@ -19,8 +19,12 @@
 package org.apache.hadoop.hive.metastore.client;
 
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.hive.metastore.HiveMetaStoreClient;
 import org.apache.hadoop.hive.metastore.IMetaStoreClient;
+
+import java.lang.reflect.InvocationHandler;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.lang.reflect.Proxy;
 
 /**
  * A synchronized wrapper for {@link IMetaStoreClient}.
@@ -28,7 +32,43 @@ import org.apache.hadoop.hive.metastore.IMetaStoreClient;
  * This should be used by multi-thread applications unless all the underlying layers are thread-safe.
  */
 public class SynchronizedMetaStoreClientProxy extends BaseMetaStoreClientProxy implements IMetaStoreClient {
+  public static SynchronizedMetaStoreClientProxy newClient(Configuration conf, IMetaStoreClient delegate) {
+    return new SynchronizedMetaStoreClientProxy(conf, delegate);
+  }
+
   public SynchronizedMetaStoreClientProxy(Configuration conf, IMetaStoreClient delegate) {
-    super(HiveMetaStoreClient.newSynchronizedClient(delegate), conf);
+    super(newSynchronizedClient(delegate), conf);
+  }
+
+  /**
+   * Creates a synchronized wrapper for any {@link IMetaStoreClient}.
+   * This may be used by multi-threaded applications until we have
+   * fixed all reentrancy bugs.
+   *
+   * @param client unsynchronized client
+   * @return synchronized client
+   */
+  public static IMetaStoreClient newSynchronizedClient(IMetaStoreClient client) {
+    return (IMetaStoreClient) Proxy.newProxyInstance(
+        ThriftHiveMetaStoreClient.class.getClassLoader(),
+        new Class[]{IMetaStoreClient.class},
+        new SynchronizedHandler(client));
+  }
+
+  private static class SynchronizedHandler implements InvocationHandler {
+    private final IMetaStoreClient client;
+
+    SynchronizedHandler(IMetaStoreClient client) {
+      this.client = client;
+    }
+
+    @Override
+    public synchronized Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
+      try {
+        return method.invoke(client, args);
+      } catch (InvocationTargetException e) {
+        throw e.getTargetException();
+      }
+    }
   }
 }
